+++ src/common/plugin/CurrencyTools.ts
