+++ src/common/plugin/ServerScores.ts
