+++ src/common/utxobased/engine/ServerStates.ts
@@ -1,16 +1,16 @@
 import { EdgeLog, EdgeTransaction } from 'edge-core-js/types'
 import { parse } from 'uri-js'
 
-import { EngineEmitter, EngineEvent } from '../../plugin/makeEngineEmitter'
-import { PluginState } from '../../plugin/pluginState'
+import { EngineEmitter, EngineEvent } from '../../plugin/EngineEmitter'
+import { PluginState } from '../../plugin/PluginState'
 import { PluginInfo } from '../../plugin/types'
 import { removeItem } from '../../plugin/utils'
 import { SafeWalletInfo } from '../keymanager/cleaners'
 import { BlockBook, makeBlockBook } from '../network/BlockBook'
-import { SubscribeAddressResponse } from '../network/BlockBookAPI'
+import { SubscribeAddressResponse } from '../network/blockbookApi'
 import Deferred from '../network/Deferred'
-import { SocketEmitter, SocketEvent } from '../network/MakeSocketEmitter'
 import { WsTask } from '../network/Socket'
+import { SocketEmitter, SocketEvent } from '../network/SocketEmitter'
 import { pushUpdate, removeIdFromQueue } from '../network/socketQueue'
 import { MAX_CONNECTIONS, NEW_CONNECTIONS } from './constants'
