+++ src/common/utxobased/network/Socket.ts
@@ -3,8 +3,8 @@ import { EdgeLog } from 'edge-core-js/types'
 
 import { removeItem } from '../../plugin/utils'
 import Deferred from './Deferred'
-import { SocketEmitter, SocketEvent } from './MakeSocketEmitter'
 import { setupWS } from './nodejsWS'
+import { SocketEmitter, SocketEvent } from './SocketEmitter'
 import { pushUpdate, removeIdFromQueue } from './socketQueue'
 import { InnerSocket, InnerSocketCallbacks, ReadyState } from './types'
 import { setupBrowser } from './windowWS'
