+++ test/common/plugin/CurrencyPlugin.fixtures/common.ts
