+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/bitcoin.ts
