+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/bitcoincash.ts
