+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/bitcoinsv.ts
