+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/digibyte.ts
