+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/feathercoin.ts
