+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/groestlcoin.ts
