+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/litecoin.ts
