+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/smartcash.ts
