+++ test/common/plugin/CurrencyPlugin.fixtures/currencies/zcoin.ts
