+++ test/common/plugin/CurrencyPlugin.fixtures/index.ts
